BuchbergerLearning

Documentation for BuchbergerLearning.

BuchbergerLearning.add_generator!

BuchbergerLearning.add_reducer!

BuchbergerLearning.buchberger

BuchbergerLearning.degree

BuchbergerLearning.find_reducers

BuchbergerLearning.iscoprime

BuchbergerLearning.ismonic

BuchbergerLearning.lc

BuchbergerLearning.li

BuchbergerLearning.lm

BuchbergerLearning.lt

BuchbergerLearning.lx

BuchbergerLearning.normalize

BuchbergerLearning.reduces

BuchbergerLearning.remove_reducer!

BuchbergerLearning.select

BuchbergerLearning.spoly

BuchbergerLearning.update!

BuchbergerLearning.add_generator! — Method

add_generator!(G::Vector{T}, R::TrieNode, f::T, o::MonomialOrdering) where {T<:MPolyRingElem}

Add $f$ to the generators $G$ and reducers $R$.

BuchbergerLearning.add_reducer! — Method

add_reducer!(node::TrieNode, f::MPolyRingElem, idx::Int, o::MonomialOrdering)

Add the reducer $f$ at index $idx$ to the trie.

BuchbergerLearning.buchberger — Method

buchberger(I::MPolyIdeal{T}; complete_reduction::Bool=true, elimination_strategy::EliminationStrategy, selection_strategy::SelectionStrategy) where {T<:MPolyRingElem}

If the internal ordering of $I$ is global, return a Gröbner basis of $I$ with respect to this ordering.

Note

The returned Gröbner basis is reduced if complete_reduction = true.

BuchbergerLearning.degree — Method

degree(f::MPolyRingElem)

Return the degree of $LM(f)$.

BuchbergerLearning.find_reducers — Method

find_reducers(trie::TrieNode, f::MPolyRingElem, o::MonomialOrdering)

Return an iterator over the indices of all polynomials in the trie reducing $f$.

BuchbergerLearning.iscoprime — Method

iscoprime(f1::T, f2::T, o::MonomialOrdering) where {T<:MPolyRingElem}

Return whether the lead monomials of $f1$ and $f2$ are coprime.

BuchbergerLearning.ismonic — Method

ismonic(f::MPolyRingElem, o::MonomialOrdering)

Return whether $f$ is monic, i.e. $LC(f) = 1$.

BuchbergerLearning.lc — Method

lc(f::MPolyRingElem, o::MonomialOrdering)

Return the leading coefficient of $f$ w.r.t. to the monomial ordering $o$.

BuchbergerLearning.li — Method

li(f::MPolyRingElem, o::MonomialOrdering)

Return the index of the leading term of $f$ w.r.t. to the monomial ordering $o$.

BuchbergerLearning.lm — Method

lm(f::MPolyRingElem, o::MonomialOrdering)

Return the leading monomial of $f$ w.r.t. to the monomial ordering $o$.

BuchbergerLearning.lt — Method

lt(f::MPolyRingElem, o::MonomialOrdering)

Return the leading term of $f$ w.r.t. to the monomial ordering $o$.

BuchbergerLearning.lx — Method

lx(f::MPolyRingElem, o::MonomialOrdering)

Return the exponent vector of the leading term of $f$ w.r.t. to the monomial ordering $o$.

BuchbergerLearning.normalize — Method

normalize(f::T, o::MonomialOrdering) where {T<:MPolyRingElem}

Normalize non-zero $f$. Return the resulting monic polynomial.

BuchbergerLearning.reduces — Method

reduces(f::T, g::T, o::MonomialOrdering) where {T<:MPolyRingElem}

Return whether $f$ reduces $g$, i.e. if $LM(f) | LM(g)$.

BuchbergerLearning.remove_reducer! — Method

add_reducer!(node::TrieNode, f::MPolyRingElem, idx::Int, o::MonomialOrdering)

Remove the reducer $f$ at index $idx$ from the trie, if it exists.

BuchbergerLearning.select — Function

select(S::SelectionStrategy, G::Vector{T}, Q::Vector{Tuple{Int,Int}}, o::MonomialOrdering) where {T<:MPolyRingElem}

Select a critical pair from $Q$ according to the provided strategy. Return the index of the selected pair in $Q$.

BuchbergerLearning.spoly — Method

spoly(f1::T, f2::T, o::MonomialOrdering) where {T<:MPolyRingElem}

Return the S-polynomial for $f1$ and $f2$.

BuchbergerLearning.update! — Function

update!(S::EliminationStrategy, Q::Vector{Tuple{Int,Int}}, G::Vector{T}, f::T, o::MonomialOrdering) where {T<:MPolyRingElem}

Update the critical pairs $Q$ when adding $f$ to $G$. Eliminate critical pairs according to the provided strategies.